Cards Project CardsDeck Class and Deck Class
CardsDeck is the cards class with the descriptions, the card deck, labels, suits, suit types, hands dealt, cards dealt per hand, and the number of hands dealt. Deck is the child class with the parameters of the parent class and will be the class that shuffles and deals the deck.